+++ src/test/java/de/stklcode/jvault/connector/model/response/MetaSecretResponseTest.java
@@ -132,6 +132,7 @@ class MetaSecretResponseTest extends AbstractModelTest<MetaSecretResponse> {
 
     private void assertSecretData(SecretResponse res) {
         assertNotNull(res, "Parsed response is NULL");
+        assertEquals(SECRET_REQUEST_ID, res.getRequestId(), "Incorrect request ID");
         assertEquals(SECRET_LEASE_ID, res.getLeaseId(), "Incorrect lease ID");
         assertEquals(SECRET_LEASE_DURATION, res.getLeaseDuration(), "Incorrect lease duration");
         assertEquals(SECRET_RENEWABLE, res.isRenewable(), "Incorrect renewable status");